pool: Fix partial retrieve for FTP MODE E
Because we extracted the range in the constructor of the Mode instance,
the range given by the door was never applied. The effect is that we
send the complete file rather than just the range requested by the
client.

This turns out to be a serious compatibility issue with ARC, although
it is unknown which versions of ARC this doesn't work with.

Target: trunk
Target: 1.9.10
Target: 1.9.9
Target: 1.9.8
Target: 1.9.5
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2645/