auth: fix NPE in Subjects class
The Subjects utility class contains several methods that rely on a private static method 
getUniquePrincipal.  This method will throw a NullPointerException if presented with a Subject 
argument that is null.

This looks to be a bug as the calling code doesn't protect against calling the method with a null 
argument.

The patch fixes getUniquePrincipal so that, if a null Subject is presented then null is returned.

Patch is initially for trunk; we could consider back-porting to earlier releases as this bug can 
prevent further investigation of some problems.

Target: trunk
Acked-by: Tigran Mkrtchyan
Patch: http://rb.dcache.org/r/2642/