pinmanager: Disabled file check in pool manager request
PinManager asks pool manager for a read pool. The default behaviour for
pool manager is to check the presence of the file on each pool that is
registered as a file location. This check requires an additional round
trip to one or more pools.

This patch disabled this check. Instead PinManager now retries in case
the pool is disabled or the file is not found. The pool is extended such
that it will clear stale location entries in the name space on attempts
to pin files that do not exist on the pool.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2652/