srmclient: use CLASSPATH environment variable rather than on the command-line
When launching dCache domains, we switched from supplying the classpath on the 
java command-line to using the CLASSPATH environment variable.  This has the 
advantage of greatly reducing the size of the command line, which may be useful 
when using normal Unix tools like "ps" or "top".

This patch does the same for the srmclient.

Patch: http://rb.dcache.org/r/2667/
Target: trunk
Acked-by: Dmitry Litvintsev