srm: Pipline locality and space token lookup for directory listing
Verbose directory listing requires a PoolManager and a SpaceManager
lookup for each directory entry. When performed sequentially, the
communication latency severily limits the throughput.

This patch introduces pipelining of the lookups. Request messages are
sent asynchronously, with an upper limit on the number of outstanding
requests to avoid overloading PoolManager and SpaceManager.

The speedup depends on the capabilities of the hardware and on
communication latency between SRM, SpaceManager and PoolManager. My
virtual machine represents the worst possible case: Limitted resources
(i.e. no room for speedup due to increased concurrency) and single
domain (i.e. low communication overhead). Yet I observed a 5-15 %
speedup of listing a directory with 378 entries. I expect the real world
speedup to be significantly higher.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>