Xrootd-door: Create classes needed for xrootd security support
This patch introduces the classes, datastructures and serialization and
deserialization needed to support pluggable xrootd authentication. The
security handler and factory introduced in this patch only provide an
accept-all authentication. However on the basis of the framework added by
this patch, further security handlers can be added to the xrootd door in
dCache.

Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Target: trunk
Patch: http://rb.dcache.org/r/2650