Xrootd-door: Add GSI authentication support
This patch adds the required authentication factory and authentication
handler for GSI support to the xrootd door. The classes implement
exchange of symmetric key and certificates plus the verification of the
respective certificate chains.

Which kind of handler is active in the xrootd door will be configurable.
Spring dependency injection is used to make the correct factory available.

Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Target: trunk
Patch: http://rb.dcache.org/r/2665

RELEASE NOTES:
Support of strong authentication has been added to the xrootd
implementation. The first supported authentication protocol is GSI.
Whether GSI authentication is active on the xrootd door or not can be
configured in the dCache configuration. Filesystem locations and refresh
options of trust anchors and the host certificate are the same as for
the SRM door.