pool: use a special type of IO queue for request processing
The main idea behind design is to decouple thread creation from
a request. There is a MoverExecutionService and PostExecutionService
which takes care about process a request and do post processing
procedure like file location registration and billing updates.
MoverExecutionService takes as an argument request and a callback -
a CompletionHandler. The CompletionHandler ( stolen from JDK7 ) has
a methods for successful and failed execution states. Current
implementation has a LegacyExecutionService which is a new
way of per-mover threat concept.

Signed-off-by: Tatjana Baranova
Signed-off-by: TIgran Mkrtchyan
Acked-By: Gerd Behrmann
Patch: http://rb.dcache.org/r/2669