libs: Upgrade JDBC, DBCP and C3P0 libraries
Upgrade to the Postgresql 9.0 version of the JDBC driver (which is
backwards compatible with Postgresql all the way back to Postgresql 7.2.
Also moved to the JDBC4 version of the driver.

Upgraded to DBCP 1.4 (don't know which version we used before).

And yes, we now use Maven to import the libraries...

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2724/