unit-tests: update FQAN tests to point to OGF document
The FQAN class has some unit-tests.  As usual, these verify correct behaviour; however, I 
had mistakenly thought that /Role=NULL was due to buggy client software.  It turns out that 
this behaviour is OK, but may go away in the future.

This path updates the unit-tests to make this clear.  It also includes a reference to the 
most recent document describing the FQAN format.

No action is needed in the unit-tests as we already check the equivalence of FQANs with 
/Role=NULL and those that have no /Role statement.

Target: trunk
Acked-by: Tanja
Patch: http://rb.dcache.org/r/2734/