pnfsmanager: fix problem when doing chmod with Chimera
After some following some red herrings, it seems that RT Ticket 5575 is actually the 
result of a bug in Chimera or ChimeraNamespaceProvider:

	http://rt.dcache.org/Ticket/Display.html?id=5575

The issue is that, in Chimera there is a coupling between a file's type (File, Directory, 
Symlink, ...) and the "mode" (as per chmod).  The file type and mode are combined as a 
single integer: the type is the high-order bits and the mode is the low-order bits.  This 
coupling was originally as a result of how the information is stored in the database but 
is now artificial constraint inside the Chimera code-base.

One side-effect of this coupling is that Chimera's Stat class setMode method changes both 
the file's type and mode at the same time; there is no method to change only the 
low-order mode.

The Chimera namespace-provider provides a method for setting file attributes.  This 
method takes the user-supplied number (e.g., 0755) as the argument to setMode.

Unfortunately, this clears the high-order bits, leaving the file with a type "0", which 
is invalid.  This gives the effect reported in the RT ticket.

The fix here is for ChimeraNameSpace provider to be more careful when using setMode; 
rather than applying the user-supplied value directly, it builds a new "mode" value by 
combining the user-supplied mode with the file's previous high-order bits.  The combined 
mode is then set, so preserving the high-order bits.

I consider this patch a work-around: the problem is actually within Chimera and future 
patches will address this issue; however, this patch is sufficient to fix the issue.  
However, it could be argued that Chimera is working as intended and the bug is more that 
Chimera doesn't provide a convenient way to do chmod.

Note that, for completeness, the patch also fixes the fileMetadata2Stat; however, this 
method is apparently dead-code.

Target: trunk
Patch: http://rb.dcache.org/r/2628/
Bug: http://rt.dcache.org/Ticket/Display.html?id=5575
Request: 1.9.5
Request: 1.9.8
Request: 1.9.9
Request: 1.9.10