scripts: fix problem with dcache script dying on Solaris machines
Ticket RT #5990 described the problem where the isRunning() shell function (in 
skel/share/lib/utils.sh) can result in the script dying unexpectedly.

The cause is that the ps code returns with a non-zero return code; since the shell 
is running with "set -e", this kills the shell process.

This patch fixes the issue by "trapping" the non-zero return code and assigning a 
value to a variable if this happens.  This succeeds, so the combined operation 
always succeeds.

Patch: http://rb.dcache.org/r/2763/
Bug: http://rt.dcache.org/Ticket/Display.html?id=5990
Acked-by: Owen Synge
Target: trunk
Request: 1.9.5
Request: 1.9.8
Request: 1.9.9
Request: 1.9.10