pool: Fix P2P cancellation
Cancelling a P2P task that was already cancelled would fail with an
IllegalStateException. The task would then not be removed from the pool
and would stay in Failed forever.

This patch fixes this flaw by adding the necessary transitions to the
Cancelling state.

The patch also wraps a number of Runnable tasks in a FireAndForgetTask
to ensure proper logging of exceptions.

Fixes a problem observed an IN2P3.

Target: trunk
Request: 1.9.10
Request: 1.9.9
Request: 1.9.8
Request: 1.9.5
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2714/