httpd: Fix pool group view
The httpd service has a view of all pool groups and the pools in each
pool group. The current implementation contains a bug in which removal
of a pool does not cause the pool to be removed from the view.

The code in question is rather complicated. I gave up fixing the bug and
instead reimplemented one of the components. The new components is
lacking some of the features of the old component. I have however not
been able to understand what those features do and they are not used in
our default configuration. The old component is still available and can
be reactivated if a site needs those features.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2739/