new dCache configure scritps made more cross platform and supporting dCache 1.7 to trunk
This application is to make dCache administrators have an easier time configuring the dCache virtual storage system especially for the first time. It also facilitates teh fresh install dCache test suites, and mayh be used for manging simple configuration tasks. It builds upon the design of YAIM and intends to be 100% backwards comopatable whiel adding the ability to write modules in python.

Patch: http://rb.dcache.org/r/2482/
Signed-off-by: Paul
Target: Trunk


Completely new infrastructure.


dCacheConfigure
~~~~~~~~~~~~~~~

This application is to make dCache administrators have an easier time configuring the dCache virtual storage system especially for the first time. It also facilitates teh fresh install dCache test suites, and mayh be used for manging simple configuration tasks. It builds upon the design of YAIM and intends to be 100% backwards comopatable whiel adding the ability to write modules in python.

Invocation
~~~~~~~~~~

dCacheConfigure is a modular application tthat takes the site-info.def formateed configuration file of the same form as YAIM and applies this to the configuration of dCache where a mapping can be atchived.

A typical invocations of dCacheConfigure.sh

dCacheConfigure.sh --site-info=/pathto/site-info.def \
                   --hostname=host.example.org \
                   --configure=org_dcache_all

The script shell wrapper around python script to facilitate deployment of dCacheConfigure. The site-info.def file stores the configuration to be supplied for configuring dCache. This can be in the from of a file as in the above example or as a set of enviroment variables coresponding to the site-info.def key value pairs with the --enviroment parameter.

The "--hostname=host.example.org" specifies the fully qualified host name as gettign a computers host name is problematic in a cross platform project such as dCache or dCacheConfigure.

The python script can launch modules specied in its configuation, the modules can then be combined to make a series of targets. The targets can be listed with the "--list-targets" parameter, all targets can then be launched using the "--configure=your_target".

Targets and Modules
~~~~~~~~~~~~~~~~~~~

Users can only invoke configuration targets or "targets" as they will be known in this document. Targets can be made up of any number of modules and are executed in the order specified in the Target, modules and targets are configuration that sbstracts the execution of configuration modules. Users can query the content of any target usign the command line option "--info-target=an_example_target" This will list the number of modules thier order a short description and thier short description.

Modules cannot be accessed directly by dCacheConfigure, they must be placed within a target and will then be executed in series in the order found within the targets. Modules provide the nessisary information to execute either python or shell functions. These python and shell functions. An example module configuration file is speified below.

<quote>
[org_dcache_info]
description = "Sets up the dCache info system"

language = "python"

# files are parsed in order searching in each directory
path_files = org_dcache_info.py

# directories are parsed in order
path_directorys = .

run = org_dcache_info_run

check = org_dcache_info_check
</quote>

The first line and section title defines the name of the module. 
The description is free text and allows the perpose of the function to be clearly undersood by the user of the module. 
The "language" species the executor shell so that modules can be writern in bash/sh or python, but it is recomended to write modules in python as its easier to make modules cross platform. 

path_directories for python scripts in the standard directory, but for bash its often improtant to source a series of files to allow a YAIM module execute, dCacheConfigure takes this YAIM convetion and makes the full path exposed to the module developer to allow easy addition previously deployed YAIM modules.

The "run" and "check" keys with values that corespond to the functions that shoudl be executed, in python they shoudl return 
"True" to indicate success, for sh scripts they shoudl set teh return value of the function to 0 to indicate success. The "check" function is executed first to check if the "run" function is needed. If the "check" function returns success the "run" function is then invoked.

By default the modules are installed as "dCacheConfigure/targets/" and "dCacheConfigure/modules/" with referance to dCacheConfigure's home directory. Simply adding files to this directory will alow further YAIM or dCacheConfigure modules to be added.

site-info.def
~~~~~~~~~~~~~

The configuration values of site-info.def are documented aleady for YAIM and dCache. 

http://trac.dcache.org/projects/dcache/wiki/manuals/site-info.def