Xrootd-door: Send subject obtained via strong authentication to gPlazma
This patch adds an authorization step to the authentication message
exchange already builtin in the xrootd door. The authentication handlers
have been modified to return a subject and some metainformation about
themselves (such as if they provide strong authentication). The strategy
to be used in authorizing the user's subject is injected via Spring based
on the configuration. 

For token based authorization, an AnonymousLoginStrategy has been
added that constructs a subject based on the door configuration.

Acked-by: Gerd Behrmann <behrmann@ndgf.org>
Target: trunk
Patch: http://rb.dcache.org/r/2723