srm: Fix remote address handling in JettyGSIConnector
Our JettyGSIConnector relies on the GSIGssSocket wrapper from JGlobus.
That class doesn't provide a value for the remote address. Hence the SRM
fails to provide the client address when it is pinning files.

This patch adds the implementation of the getRemoteAddr method in our
subclass of GSIGssSocket.

Target: trunk
Request: 1.9.10
Request: 1.9.9
Request: 1.9.8
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2779/