srm: add jetty default handler
Jetty has the concept of handlers: these are components that might (or 
might not) handle an HTTP request.

Current deployment has a single handler: a webservices handler (using 
Axis).  If a client connects and makes a request outside the context 
of the webservices handler (i.e., outside of "/srm") then the 
webservices handler doesn't handle the request.

Jetty assumes that there will always be at least one handler for each 
request.  If no handler is available for a request then an NPE is 
thrown internally.  When the NPE is thrown, it gets logged and 
connection is dropped without providing the client with any feedback.

This patch alters the handlers for jetty: it builds a chain of 
handlers consisting of two items: the webservices handler and the 
DefaultHandler.  The DefaultHandler returns a correctly formatted 404 
response with suitable HTML.

Target: trunk
Patch: http://rb.dcache.org/r/2776/
Acked-by: Gerd Behrmann