srmclient: add support for SSL transport for the srm client
This patch adds support for the client to connect to the SRM via several different transports.  
This includes GSI, SSL and unencrypted TCP.  The default is GSI.

The key part of the patch is a new enum class Transport and a utility class TransportUtil.  The 
enum class describes which communication mechanism is to be used and the utility class provides 
some handy methods for dealing with the different transports.

In the dCache configuration (dcache.conf and layout file), the srmClientTransport parameter 
adjusts which transport the SRM server uses.  The default value is GSI.  This is the transport 
used when talking to all remote SRM instances for srmCopy commands.

The patch adds the option "-transport" to the srmclient cli.  This is the transport to be used 
when connecting to the SRM instance.  The default value is GSI.

The patch also removes a couple of srmclient options (webservice_protocol, gsissl) as they didn't 
do anything and were confusingly similar to the transport option.

Note that this cannot be the complete solution for srmCopy since GSI and SSL SRM instances must 
interoperate during a transitional period.  This patch is important because it provides an SRM 
client that can talk to an SRM server using SSL.

Target: trunk
Patch: http://rb.dcache.org/r/2693/
Acked-by: Dmitry Litvintsev