PinManager: Seperate CLI and message interface from core PinManager logic
Moves the CLI implementation and message interface to separate classes.

This is the first step into removing the multitude of places in which
replies are generated. Exceptions are usually a better way to propagate
errors and this patch moves the PinManager in this direction. More work
is however needed.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2751/