PinManager: Use Spring's JDBC support
This patch updates the DAO of the PinManager to use Spring's transaction
management and Spring's JdbcTemplate. This significantly simplifies the
DAO.

We still use custom transaction demarcation. This is because the
transaction scopes of PinManager are rather complicated and cannot be
easily mapped to method scopes. This will eventually have to be cleaned
up. To reduce the amount of changes in one patch I have used the
existing init, commit and rollback calls of the DAO to wrap the Spring
transaction.

The custom JDBC connection pool is no longer used. DBCP is used directly
through a DataSource.

DB related wiring is now done with Spring. To support the Pgpass file,
the PgPass class had to be modified to make it callable from Spring.

Custom schema management is still done in DAO. This will be moved to
LiquiBase in a future patch.

Since this is part of a series of patches, there are a few fixes and 
changes not directly related to the DB change. I hope the reviewer can
abstract from this.

Target: trunk
Acked-by: Thomas Zangerl <tzangerl@pdc.kth.se>
Patch: http://rb.dcache.org/r/2752/