PinManager: Move schema management to LiquiBase
This patch moves the management of the DB schema of the PinManager to
LiquiBase. Migration from version 4 of the schema is supported. The
schema itself is not changed, but the way the schema version is tracked
changes.

Migration from previous versions of the schema is not supported. An
upgrade from such a version requires that an intermediate version of
dCache is installed before upgrading to a version with this patch.

Release Notes: Direct upgrades from dCache 1.9.1 and older are not
supported. To upgrade from such a version of dCache first upgrade to
version 1.9.5 and start the PinManager to update the DB schema.

Target: trunk
Acked-by: Thomas Zangerl <tzangerl@pdc.kth.se>
Patch: http://rb.dcache.org/r/2757/