web-admin: fix NPE on shutdown
The WebCollectorV3 creates two threads, the second with a 30-second delay.  The 
shutdown code assumes that this second thread exists and attempts to interrupt 
it.  If the second thread hasn't been created yet then this throws a 
NullPointerException.

This patch fixes this by moving the delay into the thread.  The effect is the 
same: delaying the starting of the activity of the second thread, but the 
thread is always created.

Target: trunk
Patch: http://rb.dcache.org/r/2805/
Acked-by: Gerd Behrmann