properties: factor out cleaner default properties
There are two cleaners: one for PNFS and one for Chimera.  The Chimera cleaner properties mirror 
(more or less) the PNFS cleaner properties, yet the two are defined in different places in the 
share/defaults/dcache.properties file.  In keeping with other properties, the default value for 
various properties are hidden away (inside the code, inside the batch file, etc), so it may not 
be clear what is the default behaviour.

This patch tries to combine the different properties for the PNFS and Chimera cleaners.  In 
doing this, a new defaults file is created.  The old properties in dcache.properties file are 
removed.

Where the property is shared between the two, the cleaner.properties file contains a description 
that covers both cleaners.  If a property applies to only the Chimera- or only the PNFS-cleaner 
this is noted.

The patch makes the default values explicit.  It removes the default values from within the code 
or batch files, replacing these defaults with checks that some value has been assigned.

The patch also maps the old property names to the newer "dot" style.  The older properties are 
honoured, so we maintain backwards compatibility, but are marked "(deprecated)", to suggest that 
site-admins should adjust their configuration to use the newer names.

Some of the default values differ between PNFS and Chimera.  The current patch uses the Chimera 
defaults, but includes a short note indicating the PNFS default values.

A major issue is with the cleanerRecover property.  The PNFS cleaner treats this value as a 
period in minutes and the Chimera treats this as a period in seconds.

Target: trunk
Patch: http://rb.dcache.org/r/2780/
Acked-by: Tigran Mkrtchyan