srmclient: store transport as string, add a TODO
The srmclient now understands the concept of different transports.  With the current trunk code, 
the srm client stores which transport to use as a Transport object.

Dmitry recommended storing the information as a String, to reduce the coupling between the 
client-code and the Transport object.

This patch implements that recommendation.

Patch: http://rb.dcache.org/r/2807/
Target: trunk
Acked-by: Dmitry Litvintsev