services: Don't rely on domain context for spring initialization
Before the new configuration system was introduced, we often relied on
the domain context to contain configuration parameters for Spring wired
services. We can no longer rely on this, as service specific
configuration parameters defined in the layout file are not available in
the domain context.

This patch updates those services that still relied on the domain
context to obtain configuration parameters. The parameters are now made
explicit in the batch file by providing them as arguments to the cell.

Eventually I plan to add explicit support to cells to load a Spring
application context. At that point we can make the CellShell environment
available to spring directly and we avoid the duplication of
configuration parameters.

I moved default declarations out of the Spring file and into the batch
files. They ought to be moved to property files in the defaults
directory, but I leave that task to the authors of those services.

I request that this is merged into 1.9.11, as the affected services can
currently not receive service specific configuration parameters.

Target: trunk
Request: 1.9.11
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2813/