scripts: Let 'dcache pool create' command update layout file
The dcache script had a command for creating a pool directory. Before
dCache 1.9.7 the dcache script had a second command to add the pool to a
domain, however that command was removed when the new configuration
system was introduced.

This patch adds functionality to the dcache pool create command for
updating the layout file. Thus a single command is enough to create the
pool directory and add the pool to a domain.

There is no longer a need to copy setup.temp to the pool directory, as
the only purpose was to set the pool size. The pool size is now
stored in the layout file.

Functionality is added to select the meta data backend and the lfs mode.

The command checks if a dCache runs as a non-root user and changes the
owner of the pool directories accordingly.

The file system size check now generates a warning rather than an error.
That's to support the use case in which a different, larger, file system
is mounted over the data directory.

This patch is essentially an updated subset of patch 2819.

I request this patch is merged with 1.9.11, as the previous version of
the script outputs wrong instructions on how to add the pool to a
domain.

Target: trunk
Request: 1.9.11
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2825/