bootloader: Fix compatibility with IBM JDK
The method DeprecatableProperties.stringPropertyNames assumed that
Properties.stringPropertyNames was modifiable. That is the case on the
Sun JDK, however it is not the case on the JBM JDK.

The JavaDoc of the method states that the returned set is not backed by
the Properties object. It does however not say whether the returned set
is modifiable or not - just that any modifications made are not
propagated to the Properties object.

This patch changes our implementation to create a copy of the set
instead.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2831/