boot: Include service specific parameters in getProperty shell function
The boot loader can output a shell function that serves as an oracle for
configuration values. This patch adds support for querying service
specific configuration values in addition to domain specific and global
configuration values.

The shell scripts currently do not make use of the service specific
values, but that will change in a future patch.

Property constants have been moved to a new class, Properties.

The init script has been extended with a hidden command for invoking the
boot loader directly. This is mainly intended as a debugging aid.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2828/