xrootd,pool: Avoid unnecesary reverse lookup of client address
Xrootd includes the client in the ProtocolInfo it sends to the pool. It
used to send the host name, which forces the door to perform a slow
reverse lookup. The pool would then resolve the name back to an address
(which we can only hope is the same address as we had originally).
Finally the pool would log the output of getCanonicalHostName, which
forces yet another reverse lookup. Even though this was only logged at
debug level, the reverse lookup would still be performed.


This patch removes the reverse lookup in the door. Instead the IP
address is sent. This automatically gets rid of the forward lookup in
the pool. The debug log statements are removed, as they seemed to be
leftovers from development.


I request this gets merged to stable branches for the simple reason that
the reverse lookups may be rather expensive (they caused a five second
delay on upload from the office to our production system).

Target: trunk
Request: 1.9.11
Request: 1.9.10
Request: 1.9.9
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2835/