scripts: add user process is running as in "dcache status" output
The dcache command has a command "status" that provides some basic information about the dCache 
domains that are running on the machine.  Traditionally, dCache's Java processes has always run 
as root but recent(ish) changes allowed the sys-admin to configure which user the Java process is 
to running as.

This patch adds an extra column to the dCache "status" command, showing which user the Java 
process is running under.

Note that this patch deliberately uses the 'echo' command, despite this being non-portable.  
Another patch is available in ReviewBoard that changes the necessary 'echo' commands to 'printf'.

Target: trunk
Patch: http://rb.dcache.org/r/2850/
Acked-by: Gerd Behrmann