scripts: fix column layout in dcache command
A recent patch switched how column layout was handled in the dcache command.  It used the tab 
character to separate columns.

Unfortunately there are too many shell implementations that do not adhere to POSIX semantics for 
the echo command when there are escaped characters (such as "\t" for tab).  Because of this, it 
is recommended to use printf when printing characters with escaped characters.

This patch switches those echo statements with "\t" to use printf instead of echo, thus fixing 
the column layout.

Target: trunk
Patch: http://rb.dcache.org/r/2843/
Acked-by: Karsten Schwank
Acked-by: Gerd Behrmann