scripts: show configured user if domain isn't running
Current trunk will show the process a domain is running as within the output of the 
"dcache status" command.  If the domain isn't running then the USER field is empty.

This patch adjusts the "dcache status" behaviour.  If a domain isn't running then the USER 
field contains the name of the user the domain is configured to run as.  If the 
dcache.user property hasn't been configured then the USER field contains '[whoever runs 
"dcache start"]'

Target: trunk
Patch: http://rb.dcache.org/r/2851/
Acked-by: Gerd Behrmann