boot: add ability to print configuration as a list of XML entities.
There's currently no connection between the pure XML world of the GLUE 
info-provider and the dCache configuration system.  This leads to us 
requiring the user to write their configuration in two systems: one for 
dCache in general and another for the info-provider.

This is undesirable: it's confusing having two configuration systems and 
it's apparently hard for a site-admin to configuration XML files.

This patch is the first step towards linking the dCache configuration with 
XML files. It allows the generation of an XML file containing the dCache 
properties (in global scope) as a series of XML entity definitions.  By 
including a generated file, dCache properties may be referred to using the 
parameter name as an entity name.

Patch: http://rb.dcache.org/r/2854/
Target: trunk
Acked-by: Gerd Behrmann