PoolManager: Log at a higher log-level when pool changes state.
Ticket 6013:

http://rt.dcache.org/Ticket/Update.html?id=6013

requests that when a pool is disabled that this gets logged in the PoolManager log file.  
There is some logging for this, but it is only at info level; so it appears in the pinboard 
but not the log file.

This patch changes the log-level so the message also appears in the log file.

Patch: http://rb.dcache.org/r/2860/
Acked-by: Trigran Mkrtchyan
Acked-by: Tanja Baranova
Target: trunk
Request: 1.9.11
Request: 1.9.10
Request: 1.9.9
Request: 1.9.5