chimera: add migrated chimera configuration to global context
When migrating a 1.9.5 configuration, Chimera-connection information is taken from the 
/opt/d-cache/config/chimera-config.xml file and introduced as part of the new configuration 
system.  The property assignments are written so they are local the "pnfsmanager" service; 
typically something like:

  [PnfsManagerDomain]
  [PnfsManagerDomain/PnfsManager]
  chimera.db.user = postgres

There are support scripts that need to know Chimera configuration information (database name, 
type, etc).  This requires the migrated Chimera-connection information to be present with 
global scope.

This patch changes where Chimera configuration is written.  After the patch, the information 
is written to the dcache.conf file, so it is available everywhere.

Patch: http://rb.dcache.org/r/2861/
Acked-by: Tigran Mkrtchyan
Target: trunk
Request: 1.9.11