libs: Fix log4j reference
To avoid pulling in log4j as an dependency through maven we declared
log4j to be provided. We claimed to provide version 'inifity'. The
intention was that 'infinity' lexicographically is larger than all
regular version numbers. Maven however actually tries to check for the
existence of log4j with that version, which of course fails.

This patch changes the dependency to refer to the actual version that
would otherwise be pulled in.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2875/