cells: Make MessageReply generic and implement Future
MessageReply is a dCache specific delayed reply. This patch turns it
into a generic to make it more type safe. The patch also lets
MessageReply implement the Future interface since that is essentially
what a MessageReply is (a future result that happens to be a reply).

PinManager is updated accordingly.

This change is made in preparation of updating the old PinManager to use
the messages of the new PinManager.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2871