PinManager: Replace messages with those from the new PinManager
The new pin manager uses a different set of messages to interface with
SRM. This patch updates the old pin manager to use the same messages.
Consequently we can package both the old and the new pin manager as two
different services.

The only original message kept is the PinManagerMoveMessage, since that
is needed to keep compatibility with pools. The base class,
PinManagerMessage, is however removed.

There is some terminology mismatch between the messages and the
internals of the old pin manager:

- request id is in the pin manager called srm request id
- pin id is in the pin manager called pin request id
- what the pin manager calls pin id is never exposed to the outside

The old messages did not even have a consistent use of terminology.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2873/