batch: Make services using Chimera show up in 'database ls' output
The init script got a command to list databases configured for use with
services on the given host. This patch updates the services that use
Chimera to show up in the output of that command.

While at it, I cleaned up a couple of other things in the batch files I
touched.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2878/