cells: don't fail if password file isn't present
The UserPasswords class wraps a password file, allowing queries against the file's 
content.  If the file doesn't exist then the class could provide the same information as 
if the file was empty.

The current code almost supports this; the only problem comes when the class wishes to 
check if the file has been updated.  This check requires the file to exist and fails the 
operation if it doesn't.

This patch fixes that behaviour: if the file doesn't exist then update() becomes a no-op.  
This removes a requirement when installing dCache.

It also fixes a FileInputStream leak.

Target: trunk
Patch: http://rb.dcache.org/r/2882
Acked-by: Tigran Mkrtchyan