exceptions: Drop DirNotExistsCacheException
The above exception was caught and processed in several places, however
since 1.9.4 there is no code that throws this exception. Even worse, the
Transfer class relied on this exception in the logic that creates parent
directories. Thus doors that rely on this functionality are broken
without this patch. Only the xrootd door relies on this functionality.

Target: trunk
Request: 1.9.11
Request: 1.9.10
Request: 1.9.9
Acked-by: Trigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2885/