srm: fix bug in altering 3rd-party copy requests (srmAbortFile, srmExtendFileLifeTime, srmReleaseFiles)
Patch 14704 altered the SRM's use of String as storage of SURLs to use URI.  Unfortunately 
(as discovered by FindBugs) this patch introduced a bug; after the patch, 
CopyRequest.getFileRequestBySurl uses the String.equals method against an object of type 
URI.  This will always return false.

The result is that the SRM is unable to find previous established third-party transfer 
requests.  This results in operations that somehow adjust a 3rd party copy (e.g., aborting 
a 3rd party request or extending the lifetime) will always fail.

This patch fixes the problem.  The problem is also present in 1.9.11 branch, so I request 
back-porting the patch to that branch.

Target: trunk
Request: 1.9.11
Patch: http://rb.dcache.org/r/2887
Acked-by: Dmitry Litvintsev
Acked-by: Gerd Behrmann