core: use InetSocketAddress in protocolInfo
this patch is the fist step to add a support
for IPv6. By using InetSocketAddress to identify
a client we can pic a correct interface or IP
when we send it back to the client.

currently only xroot http(webdav) and nfs movers are updated.

Acked-By: Gerd Behrmann
Patch: http://rb.dcache.org/r/2839/