cells: add command for checking whether a file exists or a cell is running
The current dCache code has a couple of places where startup behaviour is 
dependent on other factors that are difficult to establish without direct 
measurements; in particular, httpd, gPlazma and ThreadManager services.

When starting up the httpd service, the batch script attempts to run the file 
/opt/d-cache/etc/httpd.conf.  This allows sites to customise dCache httpd 
interface without modifying the batch files.  By default we don't ship dCache with 
such a file, so the attempt to read (and execute) this file fails.  Unfortunately, 
this failure results in an error message being logged.  To work-around this issue, 
we emit another error message explaining that the previous one isn't an error 
message (oh, the irony!).

There are a few services (gPlazma, ThreadManager) that, when starting up, test if 
there's already a running instance.  Currently this is done using the getcellinfo 
command.  This has the disadvantage that, if the cell isn't running, an error 
message is generated.  This is bad as the cell not running isn't an error-case; it 
only results in a different (equally valid) code-path.  What's worse is that the 
not-already-running case is by far the more commonly encountered: common 
configuration results in a spurious error message.  An example of such a spurious 
error message is:

02 Dec 2010 19:05:07 () [] file:/opt/d-cache/share//services/gplazma.batch: line 
25: Command failed ((68) not found : gPlazma)

To support these use-cases, this patch introduces the "test" command.  The command 
currently checks whether a file exists or a cell has been started but the code is 
written so other tests can be added easily.

Target: trunk
Bug: http://rt.dcache.org/Ticket/Display.html?id=5944
Patch: http://rb.dcache.org/r/2898/
Acked-by: Gerd Behrmann